% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_sim_output.R
\name{check_simulation_output}
\alias{check_simulation_output}
\title{Checks the validity of the \code{simulation_output} class.}
\usage{
check_simulation_output(object)
}
\arguments{
\item{object}{Instance of the \code{simulation_output} class.}
}
\value{
Boolean or errors
}
\description{
Checks the validity of the \code{simulation_output} class.
}
