% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_sim_output.R
\name{get_trait_data}
\alias{get_trait_data}
\title{Get trait data.}
\usage{
get_trait_data(object)
}
\arguments{
\item{object}{A \code{simulation_output} object.}
}
\value{
A \code{data.table} object.
}
\description{
Get trait data.
}
