% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make_network.R
\name{make_network}
\alias{make_network}
\title{Get network data from \code{simulation_output} objects.}
\usage{
make_network(object, weight_threshold)
}
\arguments{
\item{object}{A \code{simulation_output} object.}

\item{weight_threshold}{Weight below which edges are excluded.}
}
\value{
A \code{tidygraph} objects.
}
\description{
Get network data from \code{simulation_output} objects.
}
