% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot networks from data from \code{simulation_output} objects.}
\usage{
plot_network(g, colour_by)
}
\arguments{
\item{g}{A \code{tidygraph} object.}

\item{colour_by}{A variable name, \emph{not} as a string.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot networks from data from \code{simulation_output} objects.
}
