% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_summary.R
\docType{methods}
\name{str.simulation_output}
\alias{str.simulation_output}
\title{Show the structure of a \code{simulation_output} object.}
\usage{
\S4method{str}{simulation_output}(object)
}
\description{
Show the structure of a \code{simulation_output} object.
}
