% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_summary.R
\docType{methods}
\name{summary.simulation_output}
\alias{summary.simulation_output}
\title{Summarise a \code{simulation_output} object.}
\usage{
\S4method{summary}{simulation_output}(object)
}
\description{
Summarise a \code{simulation_output} object.
}
